Flink: new Flink sink base on the unified sink

+++ flink/v1.15/flink/src/main/java/org/apache/iceberg/flink/sink/RowDataTaskWriterFactory.java
@@ -84,7 +84,7 @@ public class RowDataTaskWriterFactory implements TaskWriterFactory<RowData> {
   }
 
   @Override
-  public void initialize(int taskId, int attemptId) {
+  public void initialize(int taskId, long attemptId) {
     this.outputFileFactory = OutputFileFactory.builderFor(table, taskId, attemptId).build();
   }
